Update pom.xml to show compiler warnings (#54)
* Update pom.xml to show compiler warnings

* JDO-817: First version removing compiler warnings

* JDO-817: removing compiler warnings Raw use og parameterzied class

* JDO-817 removing compiler warnings: unused import

* JDO-817 removing compiler warnings: declaration can have final modifier

* JDO-817 removing compiler warnings: field may be final

* JDO-817 removing compiler warnings: call toString on an array

* JDO-817 removing compiler warnings: Duplicate branch in switch

* JDO-817 removing compiler warnings: Redundant String operation

* JDO-817 removing compiler warnings: redundant type cast

* JDO-817 removing compiler warnings: unnecessary interface modifier, unnecessary semicoln

* JDO-817 removing compiler warnings: unnecessary call to toString + unnecessary conversion to String

* JDO-817 removing compiler warnings: getCompanyModelInstances methods return correct generic list

* JDO-817: corrected namoing of instance variables

* JDO-817 removing compiler warnings: QueryElementHolder becomes generic class

* JDO-817 removing compiler warnings: @SuppressWarnings for set methods in fieldtypes classes

* JDO-817 removing compiler warnings: using pm.newQuery() with class argument

* JDO-817 removing compiler warnings: SuppressWarnings when creating a query from singleStringJDOQL

* JDO-817: removing compiler warnings add profile warnings to configure compiler-plugin

* JDO-817 removing compiler warnings: replace for/while loops with enhanced for loop

* JDO-817 removing compiler warnings: empty catch block, empty finally, throw inside finally

* JDO-817 removing compiler warnings: unchecked assignment / cast

* JDO-817 removing compiler warnings: raw use of parameterized class

* JDO-817 removing compiler warnings: static variable should be qualified by type name

* JDO-817 removing compiler warnings: remove unused imports

* JDO-817 removing compiler warnings: static variable should be qualified by type name

* JDO-817 removing compiler warnings: serializable class has no definition of serialVersionUID

* JDO-817 removing compiler warnings: miscellaneous warnings

* JDO-817 removing compiler warnings: rawtypes warning

* JDO-817 removing compiler warnings: Replace StringBuffer by StringBuilder

* JDO-817 removing compiler warnings: type can be replaced by <>

* JDO-817 removing compiler warnings: identical catch branches in try

* JDO-817 removing compiler warnings: try finally replaceable with try with resources

* JDO-817 removing compiler warnings: if can be replaced with switch;  redundant throws clause

* JDO-817 removing compiler warnings: use -Xlint:all,-try,-rawtypes as default

* JDO-817 removing compiler warnings: added a new paragraph about wraings to the readme

* JDO-817 removing compiler warnings: "raw use of parameterized class" in api submodule

* JDO-817 removing compiler warnings: method close w/o throws clause in Query and Extent

* JDO-817 removing compiler warnings: „redundant cast“ in api submodule

* DO-817 removing compiler warnings: „serializable class has no definition of serialVersionUID“ in api submodule

* JDO-817 removing compiler warnings: "raw use of parameterized class" in api submodule

* JDO-817 removing compiler warnings: "raw use of parameterized class" in api submodule

* JDO-817 removing compiler warnings: „static method declared final“ in api submodule

* JDO-817 removing compiler warnings: „while loop replaceable with enhanced for loop“ in api submodule

* JDO-817 removing compiler warnings: „anonymous can be replaced by lambda“ in api submodule

* JDO-817 removing compiler warnings: „field can be final" in api submodule

* JDO-817: Change pom version to 3.3-SNAPSHOT

+++ api/src/test/java/javax/jdo/identity/IntIdentityTest.java
@@ -43,29 +43,29 @@ public class IntIdentityTest extends SingleFieldIdentityTest {
     }
     
     public void testConstructor() {
-        IntIdentity c1 = new IntIdentity(Object.class, (int)1);
-        IntIdentity c2 = new IntIdentity(Object.class, (int)1);
-        IntIdentity c3 = new IntIdentity(Object.class, (int)2);
+        IntIdentity c1 = new IntIdentity(Object.class, 1);
+        IntIdentity c2 = new IntIdentity(Object.class, 1);
+        IntIdentity c3 = new IntIdentity(Object.class, 2);
         assertEquals("Equal IntIdentity instances compare not equal.", c1, c2);
         assertFalse ("Not equal IntIdentity instances compare equal", c1.equals(c3));
     }
 
     public void testIntegerConstructor() {
-        IntIdentity c1 = new IntIdentity(Object.class, (int)1);
-        IntIdentity c2 = new IntIdentity(Object.class, Integer.valueOf((int)1));
-        IntIdentity c3 = new IntIdentity(Object.class, Integer.valueOf((int)2));
+        IntIdentity c1 = new IntIdentity(Object.class, 1);
+        IntIdentity c2 = new IntIdentity(Object.class, Integer.valueOf(1));
+        IntIdentity c3 = new IntIdentity(Object.class, Integer.valueOf(2));
         assertEquals ("Equal intIdentity instances compare not equal.", c1, c2);
         assertFalse ("Not equal IntIdentity instances compare equal", c1.equals(c3));
     }
 
     public void testToStringConstructor() {
-        IntIdentity c1 = new IntIdentity(Object.class, (int)1);
+        IntIdentity c1 = new IntIdentity(Object.class, 1);
         IntIdentity c2 = new IntIdentity(Object.class, c1.toString());
         assertEquals ("Equal IntIdentity instances compare not equal.", c1, c2);
     }
 
     public void testStringConstructor() {
-        IntIdentity c1 = new IntIdentity(Object.class, (int)1);
+        IntIdentity c1 = new IntIdentity(Object.class, 1);
         IntIdentity c2 = new IntIdentity(Object.class, "1");
         IntIdentity c3 = new IntIdentity(Object.class, "2");
         assertEquals ("Equal IntIdentity instances compare not equal.", c1, c2);
@@ -82,7 +82,7 @@ public class IntIdentityTest extends SingleFieldIdentityTest {
     }
     
     public void testSerialized() {
-        IntIdentity c1 = new IntIdentity(Object.class, (int)1);
+        IntIdentity c1 = new IntIdentity(Object.class, 1);
         IntIdentity c2 = new IntIdentity(Object.class, "1");
         IntIdentity c3 = new IntIdentity(Object.class, "2");
         Object[] scis = writeReadSerialized(new Object[] {c1, c2, c3});
